renamed mtdict to mtdata, added starting_rho_estimate
